+++ lib/src/shortener/date_format.dart
@@ -1,12 +1,11 @@
 import 'package:intl/intl.dart';
 
 extension StringToDateTime on String {
-  String toDateTime() => length > 0
-      ? DateFormat('dd MMM yyyy | hh:mm aaa')
-          .format(DateTime.tryParse(this)?.toLocal() ??
-              DateTime.fromMillisecondsSinceEpoch(
-                int.parse(this),
-              ).toLocal())
+  String toDateTime({String format = 'dd MMM yyyy | hh:mm aaa'}) => length > 0
+      ? DateFormat(format).format(DateTime.tryParse(this)?.toLocal() ??
+          DateTime.fromMillisecondsSinceEpoch(
+            int.parse(this),
+          ).toLocal())
       : this;
   String toTime() => length > 0
       ? DateFormat('hh:mm aaa').format(DateTime.tryParse(this)?.toLocal() ??